pyproject.toml
[project]
name = "nonebot-plugin-mcpic"
version = "0.1.2"
dependencies = [
    "httpx>=0.26.0",
    "nonebot2>=2.1.3",
    "nonebot-adapter-onebot>=2.3.1",
]
requires-python = ">=3.8"
authors = [{ name = "wlm3201", email = "wlm3201@qq.com" }]
description = "随机发送MC建筑图片"
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["mcpic"]
classifiers = []

[project.urls]
Homepage = "https://github.com/wlm3201/nonebot-plugin-mcpic"
